Add incident AL-Birmingham-2
Added incident AL-Birmingham-2, the arrest of The Root journalist Michael Harriot. There is a slight discrepancy between the coverage of the incident in AL.com and the way Harriot describes it. (See description. There is video, but I can't make out exactly what was said from it. My working theory is that Harriot (a respected writer at a major publication) is probably telling the truth and AL.com (a smaller publication that reads rather like my school newspaper) probably screwed up.)