Added placeholder tag for az-phoenix-3
I know this isn't a valid tag; considering the mass arrests and how the
wrists of most arrestees are positioned I am almost certain zie ties
were used. But the videos are all of too poor quality and too dark to
tell for sure, and no text confirms the use of zip ties.